edit spec test
